requirements.txt
annotated-types==0.7.0
beautifulsoup4==4.12.3
click==8.1.7
cloudscraper==1.2.71
openai==1.34.0
pandas==2.2.2
pytest==8.2.2
requests==2.32.3
urllib3==2.2.2
python-dotenv==1.0.1
setuptools~=65.5.1